# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chatee API",
    "version": "1.0.0",
    "description": "API used for chatee application"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/auth/signup/": {
      "post": {
        "description": "Sign up user for the application",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "description": "Sign up body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSchema"
              },
              "example": {
                "email": "test.user@mail.com",
                "password": "test_test"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been signed up",
            "content": {
              "appication/json": {
                "example": {
                  "status": 200,
                  "logged": true,
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "message": "Sign up successfull"
                }
              }
            }
          },
          "400": {
            "description": "Sign up failed",
            "content": {
              "application/json": {
                "example": {
                  "status": 400,
                  "logged": false,
                  "message": "Email already exists"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login/": {
      "post": {
        "description": "Login user to the application",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "description": "login body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSchema"
              },
              "example": {
                "email": "test.user@mail.com",
                "password": "test_test"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been logged in",
            "content": {
              "appication/json": {
                "example": {
                  "status": 200,
                  "logged": true,
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
                  "message": "Login successfull"
                }
              }
            }
          },
          "401": {
            "description": "Not logged, invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "status": 401,
                  "logged": false,
                  "message": "Invalid credentials"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/messages/": {
      "post": {
        "description": "Create a new message",
        "tags": [
          "messages"
        ],
        "requestBody": {
          "description": "Message body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSchema"
              },
              "example": {
                "message": {
                  "type": "text",
                  "message": "asdf hi"
                },
                "chatid": "1234chatid"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message created",
            "content": {
              "appication/json": {
                "example": {
                  "status": 200,
                  "_id": "messageid1234",
                  "author": "steve@apple.com",
                  "type": "text",
                  "message": "asdf hi",
                  "createdDate": "2019-11-02T02:22:25.000Z",
                  "editDate": "2019-11-02T02:22:25.000Z"
                }
              }
            }
          }
        }
      },
      "patch": {
        "description": "Edit a message",
        "tags": [
          "messages"
        ],
        "requestBody": {
          "description": "Message body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageSchema"
              },
              "example": {
                "message": {
                  "type": "text",
                  "message": "asdf hi"
                },
                "messageid": "messageid1234"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message modified",
            "content": {
              "appication/json": {
                "example": {
                  "status": 200,
                  "_id": "messageid1234",
                  "author": "steve@apple.com",
                  "type": "text",
                  "message": "asdf hi",
                  "createdDate": "2019-11-02T02:22:25.000Z",
                  "editDate": "2019-11-02T02:22:25.000Z"
                }
              }
            }
          }
        }
      }
    },
    "/v1/messages/{id}": {
      "delete": {
        "description": "Delete a message",
        "tags": [
          "messages"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "the unique messageid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message modified",
            "content": {
              "appication/json": {
                "example": {
                  "status": 200,
                  "_id": "messageid1234",
                  "author": "steve@apple.com",
                  "type": "text",
                  "message": "asdf hi",
                  "createdDate": "2019-11-02T02:22:25.000Z",
                  "editDate": "2019-11-02T02:22:25.000Z"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageSchema": {
        "required": [
          "author"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "createdDate": {
            "type": "array"
          },
          "editDate": {
            "type": "string"
          }
        }
      },
      "UserSchema": {
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "profile": {
            "type": "object"
          },
          "password": {
            "type": "string"
          },
          "chats": {
            "type": "array"
          },
          "friends": {
            "type": "array"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    }
  },
...
}